[#4619] Rename the automations recipe to vertical slices and add event models
The concept the framework leans on here is Vertical Slice Architecture, so
the recipe and its guide page now lead with that name rather than "automation
slices", which is the Event Modelling term for the same shape. Both the guide
and the sample keep a pointer back to automation slices so readers who model
that way still recognise it.

The rename runs end to end: the saga-recipes package and its paired
saga.recipe selector value become verticalslices, the docs sample and its
include paths follow, and the guide page, navigation, cross-references and the
saga migration path are updated to match.

The saga guide index also gains a "Saga or process manager" section drawing
the classical distinction (linear event-to-command matching versus stateful
branching) and tying it to the state question the guide answers. Two prooph
board event models are added: the whole bike rental payment process on the
index page, and the payment deadline projection-and-sweep on the deadlines
page.

(cherry picked from commit e227d1c84c010b098619553a735eb43558d9e16c)
