+++ examples/saga-recipes/pom.xml
@@ -30,7 +30,7 @@
     <description>
         Four approaches to implementing an Axon Framework 4 Saga in Axon Framework 5, applied to the bike-rental
         payment process. Demonstrates process state held in a repository, derived through @InjectEntity, event-sourced
-        from the process's own events, and decomposed into automation slices.
+        from the process's own events, and decomposed into vertical slices.
     </description>
 
     <properties>
